C++ element definitions now support any kind of arithmetic types (#11)
Updated static list parser.
Now it can have same syntax as any script may have. E.g.:
- `{1,2,3}` lua style list;
- `[1,2,3]` js style list;
- `(a,b,c)` python tuple;

Old syntax without brackets is also supported.

Additionally fixed several issues in tests, improved test coverage,
added tests for parser, field validation.

Found a problem with `required` field not working for fields that are
missing from the xml node (only if field parsing failed we were getting
an error).

+++ src/lua/script.cpp
@@ -216,7 +216,7 @@ extern "C" {
 
       virtual void keys(ObjectKeys& res);
 
-      virtual int readInt() {
+      virtual long readInt() {
         StackGuard g(mLuaState);
         unwrap();
         return lua_tointeger(mLuaState, -1);
@@ -251,7 +251,7 @@ extern "C" {
         assign();
       }
 
-      virtual void setInteger(unsigned long value)
+      virtual void setInteger(long value)
       {
         lua_pushinteger(mLuaState, value);
         assign();
